+++ src/Bunker.h
@@ -11,7 +11,7 @@ public:
 	void position(int x, int y);
 	bool isdraw;
 	void drawing();
-	void rotation(int x);
+	void rotation(double x);
 private:
 	sf::ConvexShape bunker;
 	int health = 100, turrets;
