+++ src/Fuel.h
@@ -9,7 +9,7 @@ public:
 	void draw(sf::RenderWindow& w);
 	void settype(int a);
 	void position(int x, int y);
-	void rotation(int x);
+	void rotation(double x);
 	bool isdraw;
 private:
 	sf::ConvexShape benza;
